% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapButte}
\alias{bootstrapButte}
\title{Bootstrap for the CI of timing estimates}
\usage{
bootstrapButte(
  eventOrdering,
  B,
  type = c("parametric", "bootstrap"),
  pi,
  x,
  m,
  call
)
}
\arguments{
\item{eventOrdering}{is the list returned by Butte}

\item{B}{is the number of bootstrapping samples}

\item{type}{"bootstrap" for resample the data; or "parametric" to simulate randomized data}

\item{pi}{is the estimate of pi you want to use for bootstrapping}

\item{x}{the data to bootstrap}

\item{m}{the data to bootstrap}

\item{call}{list of original argument settings used in calling Butte}
}
\value{
timing estimates of bootstrapped data
}
\description{
the corresponding SCNA evolution.
}
