% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butte.R
\name{Butte}
\alias{Butte}
\title{Calculate timing of SCNA}
\usage{
Butte(
  x,
  m,
  history,
  nt,
  nb,
  qmethod = c("fullMLE", "partialMLE"),
  type = c("identifiable", "butte"),
  seqError = 0,
  bootstrapCI = NULL,
  B = 500,
  CILevel = 0.9,
  purity = 1,
  verbose = TRUE,
  returnAssignments = FALSE,
  minMutations = 10,
  init = NULL,
  maxiter = 100,
  tol = 1e-04,
  mutationId = 1:length(x),
  ...
)
}
\arguments{
\item{x}{vector of number of reads supporting SSNVs}

\item{m}{vector of total read depth for SSNVs}

\item{history}{list of possible histories returned by cnmutHistory}

\item{nt}{total copy number}

\item{nb}{copy number of the minor allele}

\item{qmethod}{suggest to use fullMLE, which is more accurate}

\item{type}{choose "identifiable" for SCNA 2:0 3:1 3:0 4:1, and "butte" otherwise}

\item{bootstrapCI}{set to "bootstrap" if non-parametric; or "parametric"}
}
\value{
A list of possible matrices
}
\description{
This function estimates the timing of SCNAs for the following two scenarios
for 2:0 3:1 3:0 4:1, where an identifiable history matrix is available, 
this function returns the time estimate and confidence interval for each stage
For SCNAs with multiple or non-indentifible matrix, this function instead
returns the bounds (lower and upper) of time period for the last stage of
the corresponding SCNA evolution.
}
