% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmut.R
\name{cnmutData}
\alias{cnmutData}
\title{Generate the SCNA SSNV input for running butte}
\usage{
cnmutData(scnaFile, ssnvFile, skipchunk = 19)
}
\arguments{
\item{scnaFile}{the SCNA segmentation file}

\item{ssnvFile}{the SSNV file}

\item{skipchunk}{segments with number of data points (probes) no more than this number will be skipped, set it to zero if not skipped}
}
\value{
list of data input for running butte
}
\description{
Four elements will be generated in the output list
merged CNA segmentation;
SSNV data.frame;
cnvHits (index in CNA file, overlapping ssnv)
snvHits (index in SSNV file, overlapping with cnv)
}
