% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQ.R
\name{.estimateQ}
\alias{.estimateQ}
\title{EM algorithm to calculate q}
\usage{
.estimateQ(
  x,
  m,
  alleleSet,
  alleleFreq = NULL,
  history,
  type = c("identifiable", "butte"),
  init = NULL,
  maxiter = 100,
  tol = 1e-04,
  xGreaterZero = TRUE,
  useGradient = FALSE
)
}
\arguments{
\item{x}{vector of number of reads supporting SSNVs}

\item{m}{vector of total read depth for SSNVs}

\item{alleleSet}{vector of possible allele frequencies}

\item{alleleFreq}{the observed allele frequencies}

\item{history}{a list of possible evolutionary history matrices, see also function "cnmutHistory"}

\item{type}{set it to be either "identifiable" or "butte"}

\item{init}{the initial values of vector q (probability of a randomly acquired SSNV having each allele state)}

\item{maxiter}{maximum number of iterations in calculation q}

\item{tol}{the tolerance in the convergence of q}

\item{xGreaterZero}{determines whether the likelihood will account for the fact that only observe SSNVs with mutant read count x[i] > 0}
}
\value{
A list of possible matrices
}
\description{
Estimate the proportion the proportions of SSNVs at each allele state
from sequencing data.
}
