% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmut.R
\name{mergeCNA}
\alias{mergeCNA}
\title{Merge the CNA by jumping (neglecting) small segments}
\usage{
mergeCNA(cnFile, skipchunk = 19, correctMale = FALSE)
}
\arguments{
\item{cnFile}{the SCNA segmentation file}

\item{skipchunk}{segments with number of data points (probes) no more than this number will be skipped, set it to zero if not skipped}

\item{correctMale}{logical, whether or not divide by 2 for the X chromosome (testing)}
}
\value{
data frame of the merged CNA segmentation
}
\description{
When CNA segmentation contains many small segments, one may want to
merge the two neighboring segments by skipping the small segment.
}
