% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmut.R
\name{scnaTiming}
\alias{scnaTiming}
\title{estimate Timing for each SCNA segment}
\usage{
scnaTiming(
  scnaFile,
  ssnvFile,
  sn,
  outname,
  public = FALSE,
  pubOrSub = "pubOrSub",
  skipchunk = 19,
  mmut = 10,
  qmethod = "fullMLE",
  B = 100
)
}
\arguments{
\item{scnaFile}{the SCNA segmentation file}

\item{ssnvFile}{the SSNV file}

\item{sn}{sample name}

\item{outname}{output timing file name (including path)}

\item{public}{somatic timeline is focused on public mutations (across multi-samples) or not (otherwise, clonal variants in the specific sample)}

\item{pubOrSub}{the colname for column indicating if the mutation is public or not}

\item{skipchunk}{segments with number of data points (probes) no more than this number will be skipped}

\item{mmut}{minimum number of mutations for running timing analysis}

\item{qmethod}{the method for estimating q (probabilities of a randomly acquired mutation having allele state of aj/Nt)}
}
\value{
list: timing result; timing table (for visualization) and merged CNA data frame
}
\description{
estimate Timing for each SCNA segment
}
