% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmut.R
\name{ssnvInput}
\alias{ssnvInput}
\title{Reading and sorting ssnvFile}
\usage{
ssnvInput(snvFile)
}
\arguments{
\item{ssnvFile}{the SSNV file}
}
\value{
sorted ssnv data frame
}
\description{
Reading and sorting ssnvFile
}
