% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vafUtil.R
\name{vafEst}
\alias{vafEst}
\title{Return the allele frequencies of SSNVs for each allele state}
\usage{
vafEst(ncmut = "all", nt, nb, pu, pa = 1)
}
\arguments{
\item{ncmut}{the number of copies for the mutation, default print for "all"}

\item{nt}{total copy number}

\item{nb}{copy number of minor allele}

\item{pu}{tumor purity}

\item{pa}{prevalence of the SCNA WITHIN the tumor content, default 1 (clonal)}
}
\value{
a named vector of expected allele frequency at each allele state
}
\description{
Given a SCNA configuration Nt (total copy) and Nb (minor copy),
as well as the tumor purity and prevalence of the SCNA,
this function returns the expected allele frequencies for possible 
allele states.
}
