% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butte.R
\name{Butte}
\alias{Butte}
\title{Calculate timing of SCNA}
\usage{
Butte(
  x,
  m,
  history,
  nt,
  nb,
  qmethod = c("fullMLE", "partialMLE"),
  type = c("identifiable", "butte"),
  seqError = 0,
  bootstrapCI = NULL,
  B = 500,
  CILevel = 0.9,
  purity = 1,
  verbose = TRUE,
  returnAssignments = TRUE,
  minMutations = 10,
  init = NULL,
  maxiter = 100,
  tol = 1e-04,
  mutationId = 1:length(x),
  ...
)
}
\arguments{
\item{x}{vector of number of reads supporting SSNVs}

\item{m}{vector of total read depth for SSNVs}

\item{history}{a list of possible evolutionary history matrices, see also function "cnmutHistory"}

\item{nt}{total copy number}

\item{nb}{copy number of the minor allele}

\item{qmethod}{suggest to use fullMLE, which is more accurate}

\item{type}{set it to be either "identifiable" or "butte", or leave it unset}

\item{seqError}{sequencing errors}

\item{bootstrapCI}{set to "bootstrap" if non-parametric; or "parametric". This specify the confidence interval method. NULL if do not want to calculate CI}

\item{B}{the number of bootstrapping samples}

\item{CILevel}{confidence interval level}

\item{purity}{tumor sample purity}

\item{verbose}{logical. Turn on/off additional warnings.}

\item{returnAssignments}{logical. Whether to return the probabilistic assignments of SSNVs to allele states generated by the EM algorithm, as well as the SSNV read counts (total depth and depth of the mutant allele).}

\item{minMutations}{minimum number of SSNVs required for timing analysis}

\item{init}{the initial value of vector q passed to function .estimateQ}

\item{maxiter}{maximum number of iterations in EM algorithm when calculating the allele state distribution}

\item{tol}{the tolerance in the convergence of q}

\item{mutationId}{a vector of mutation IDs. The default is 1:length of x.}
}
\value{
A list of possible matrices
}
\description{
This function estimates the timing of SCNAs for the following two scenarios.
For 2:0 3:1 3:0 4:1, where an identifiable history matrix is available, 
this function returns the time estimate and confidence interval for each stage
For SCNAs with multiple or non-indentifible matrix, this function instead
returns the bounds (lower and upper) of time period for the last stage of
the corresponding SCNA evolution.
}
