% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQ.R
\name{.estimateQ}
\alias{.estimateQ}
\title{EM algorithm to calculate \hat{q}}
\usage{
.estimateQ(
  x,
  m,
  alleleSet,
  alleleFreq = NULL,
  history,
  type = c("identifiable", "butte"),
  init = NULL,
  maxiter = 100,
  tol = 1e-04,
  xGreaterZero = TRUE,
  useGradient = FALSE
)
}
\arguments{
\item{x}{vector of number of reads supporting SSNVs}

\item{m}{vector of total read depth for SSNVs}

\item{xGreaterZero}{determines whether the likelihood will account for the fact that only observe X[i]>0}
}
\value{
A list of possible matrices
}
\description{
Estimate the proportion the proportions of SSNVs at each allele state
from sequencing data.
}
