% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butte.R
\name{.lpbounds}
\alias{.lpbounds}
\title{Use linear programming to solve the bounds of Ta}
\usage{
.lpbounds(q, possible_histories)
}
\arguments{
\item{q}{q estimated from data}

\item{possible_histories}{matrices of possible SCNA-SSNV histories}
}
\value{
the lower and upper bounds of the time duration for the last stage
}
\description{
Yunong: add some description here
}
